backend/requirements.txt
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
